+++ CHANGELOG.md
@@ -6,9 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Translation to `SKUSpecificationField` and `SKUSpecificationValue`.
 
 ## [0.8.0] - 2020-06-17
-
 ### Added
 - Add `suggestions`, `correction`, `banners`, and `redirect`.